# docs/swagger.yaml
basePath: /v1
definitions:
...
  main.TeacherRegisterPayload:
    properties:
      email:
        type: string
      first_name:
        maxLength: 72
        type: string
      hire_date:
        type: string
      last_name:
        maxLength: 72
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      phone_number:
        type: string
      subject:
        maxLength: 128
        type: string
    required:
    - email
    - first_name
    - hire_date
    - last_name
    - password
    - phone_number
    - subject
    type: object
...
  main.UpdateTeacherPayload:
    properties:
      email:
        type: string
      first_name:
        maxLength: 72
        type: string
      hire_date:
        type: string
      last_name:
        maxLength: 72
        type: string
      phone_number:
        type: string
      subject:
        maxLength: 128
        type: string
    type: object
...
  store.Teacher:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      hire_date:
        type: string
      id:
        type: integer
      last_name:
        type: string
      phone_number:
        type: string
      subject:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for ClassNama, a school management platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ClassNama
paths:
...
  /teachers:
    get:
      operationId: getTeachers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Teacher'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get all teachers
      tags:
      - Teachers
    post:
      consumes:
      - application/json
      description: Only Execs with manager/admin roles can create new Teachers
      parameters:
      - description: Teacher registration payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.TeacherRegisterPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Returns the created Teacher
          schema:
            $ref: '#/definitions/store.Teacher'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new Teacher
      tags:
      - Teachers
  /teachers/{teacherID}:
    delete:
      operationId: deleteTeacher
      parameters:
      - description: Teacher ID
        in: path
        name: teacherID
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete a teacher
      tags:
      - Teachers
    get:
      operationId: getTeacher
      parameters:
      - description: Teacher ID
        in: path
        name: teacherID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Teacher'
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get a teacher by ID
      tags:
      - Teachers
    patch:
      consumes:
      - application/json
      operationId: updateTeacher
      parameters:
      - description: Teacher ID
        in: path
        name: teacherID
        required: true
        type: integer
      - description: Teacher update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.UpdateTeacherPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Teacher'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "409":
          description: Conflict
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update a teacher
      tags:
      - Teachers
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"